feat: <5단계> 테스트하기 어려운 코드 의존관계를 상위로 이동
