+++ src/test/java/study/step3/CarRacingTest.java
@@ -45,5 +45,15 @@ class CarRacingTest {
         assertThat(carRacing.findWinnerNames(cars, position)).isEqualTo(carNames);
     }
 
+    @Test
+    @DisplayName("0~9 사이 무작위 값을 반환한다.")
+    public void random_함수_test() throws Exception {
+        for (int i = 0; i < 10; i++) {
+            assertThat(carRacing.getRandomInt())
+                    .isGreaterThan(-1)
+                    .isLessThan(10);
+        }
+    }
+
 
 }
